Moved the KeyBindings project to LegacyKeyBindings to accommodate a new KeyBindings implementation.
