+++ Applications/Spire/CMakeLists.txt
@@ -95,7 +95,7 @@ add_subdirectory(Config/Catalog)
 add_subdirectory(Config/Charting)
 add_subdirectory(Config/Dashboard)
 add_subdirectory(Config/InputWidgets)
-add_subdirectory(Config/KeyBindings)
+add_subdirectory(Config/LegacyKeyBindings)
 add_subdirectory(Config/LegacyResources)
 add_subdirectory(Config/LegacyUI)
 add_subdirectory(Config/Library)
