+++ Applications/Spire/Config/Canvas/CMakeLists.txt
@@ -113,7 +113,7 @@ install(TARGETS Canvas CONFIGURATIONS Release
 file(GLOB test_source_files ${SPIRE_SOURCE_PATH}/CanvasTests/*.cpp)
 add_executable(CanvasTests ${test_source_files})
 target_link_libraries(CanvasTests Async Dashboard Charting PortfolioViewer
-  Utilities RiskTimer AccountViewer KeyBindings Blotter Catalog BookView
+  Utilities RiskTimer AccountViewer LegacyKeyBindings Blotter Catalog BookView
   TimeAndSales OrderImbalanceIndicator InputWidgets CanvasView Canvas
   LegacyResources LegacyUI
   debug ${CRYPTOPP_LIBRARY_DEBUG_PATH}
