+++ Applications/Spire/Config/LegacyKeyBindings/CMakeLists.txt
@@ -1,7 +1,7 @@
 find_package(Qt5Widgets REQUIRED)
-file(GLOB header_files ${SPIRE_INCLUDE_PATH}/Spire/KeyBindings/*.hpp)
-file(GLOB source_files ${SPIRE_SOURCE_PATH}/KeyBindings/*.cpp)
-file(GLOB form_files ${SPIRE_SOURCE_PATH}/KeyBindings/*.ui)
+file(GLOB header_files ${SPIRE_INCLUDE_PATH}/Spire/LegacyKeyBindings/*.hpp)
+file(GLOB source_files ${SPIRE_SOURCE_PATH}/LegacyKeyBindings/*.cpp)
+file(GLOB form_files ${SPIRE_SOURCE_PATH}/LegacyKeyBindings/*.ui)
 qt5_wrap_ui(generated_header_files ${form_files})
 find_cpp_qobject(qobject_source_files ${header_files})
 qt5_wrap_cpp(generated_source_files ${qobject_source_files}
@@ -10,13 +10,13 @@ source_group("Form Files" FILES ${form_files})
 source_group("Generated Files\\Header Files" FILES ${generated_header_files})
 source_group("Generated Files\\Source Files" FILES ${generated_source_files})
 include_directories(${CMAKE_CURRENT_BINARY_DIR})
-add_library(KeyBindings ${header_files} ${source_files}
+add_library(LegacyKeyBindings ${header_files} ${source_files}
   ${generated_header_files} ${generated_source_files})
 set_source_files_properties(${header_files} PROPERTIES HEADER_FILE_ONLY TRUE)
-set_target_properties(KeyBindings PROPERTIES
+set_target_properties(LegacyKeyBindings PROPERTIES
   STATIC_LIBRARY_FLAGS_RELEASE "${CMAKE_LIBRARY_FLAGS}"
-  OUTPUT_NAME spire_key_bindings)
-install(TARGETS KeyBindings CONFIGURATIONS Debug
+  OUTPUT_NAME spire_legacy_key_bindings)
+install(TARGETS LegacyKeyBindings CONFIGURATIONS Debug
   DESTINATION ${SPIRE_LIBRARY_INSTALL_DIRECTORY}/Debug)
-install(TARGETS KeyBindings CONFIGURATIONS Release
+install(TARGETS LegacyKeyBindings CONFIGURATIONS Release
   DESTINATION ${SPIRE_LIBRARY_INSTALL_DIRECTORY}/Release)
