+++ Applications/Spire/Include/Spire/LegacyUI/EnvironmentSettings.hpp
@@ -8,8 +8,8 @@
 #include "Nexus/Definitions/RegionMap.hpp"
 #include "Spire/BookView/BookViewProperties.hpp"
 #include "Spire/Dashboard/SavedDashboards.hpp"
-#include "Spire/KeyBindings/InteractionsProperties.hpp"
-#include "Spire/KeyBindings/KeyBindings.hpp"
+#include "Spire/LegacyKeyBindings/InteractionsProperties.hpp"
+#include "Spire/LegacyKeyBindings/KeyBindings.hpp"
 #include "Spire/LegacyUI/LegacyUI.hpp"
 #include "Spire/OrderImbalanceIndicator/OrderImbalanceIndicatorProperties.hpp"
 #include "Spire/PortfolioViewer/PortfolioViewerProperties.hpp"
