+++ Applications/Spire/Include/Spire/LegacyUI/MaxFloorSpinBox.hpp
@@ -5,7 +5,7 @@
 #include <QSpinBox>
 #include "Nexus/Definitions/Security.hpp"
 #include "Spire/Canvas/Canvas.hpp"
-#include "Spire/KeyBindings/KeyBindings.hpp"
+#include "Spire/LegacyKeyBindings/KeyBindings.hpp"
 #include "Spire/LegacyUI/LegacyUI.hpp"
 #include "Spire/Spire/Spire.hpp"
