+++ Applications/Spire/Include/Spire/LegacyUI/MoneySpinBox.hpp
@@ -5,7 +5,7 @@
 #include <QDoubleSpinBox>
 #include "Nexus/Definitions/Security.hpp"
 #include "Spire/Canvas/Canvas.hpp"
-#include "Spire/KeyBindings/KeyBindings.hpp"
+#include "Spire/LegacyKeyBindings/KeyBindings.hpp"
 #include "Spire/Spire/Spire.hpp"
 
 namespace Spire {
