+++ Applications/Spire/Source/CanvasView/OrderTaskView.cpp
@@ -20,7 +20,7 @@
 #include "Spire/CanvasView/CondensedCanvasWidget.hpp"
 #include "Spire/CanvasView/CanvasNodeNotVisibleException.hpp"
 #include "Spire/CanvasView/CondensedCanvasWidget.hpp"
-#include "Spire/KeyBindings/KeyBindings.hpp"
+#include "Spire/LegacyKeyBindings/KeyBindings.hpp"
 #include "Spire/LegacyUI/UserProfile.hpp"
 
 using namespace Beam;
