+++ Applications/Spire/Source/LegacyKeyBindings/HotkeyOverride.cpp
@@ -1,4 +1,4 @@
-#include "Spire/KeyBindings/HotkeyOverride.hpp"
+#include "Spire/LegacyKeyBindings/HotkeyOverride.hpp"
 #ifdef _MSC_VER
   #include <windows.h>
 #endif
