+++ Applications/Spire/Source/LegacyKeyBindings/SimplifiedKeyBindingsDialog.ui
@@ -340,7 +340,7 @@
   <customwidget>
    <class>Spire::InteractionsWidget</class>
    <extends>QWidget</extends>
-   <header>Spire/KeyBindings/InteractionsWidget.hpp</header>
+   <header>Spire/LegacyKeyBindings/InteractionsWidget.hpp</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
