+++ Applications/Spire/Source/LegacyKeyBindings/TaskKeyBindingsDialog.ui
