+++ Applications/Spire/Source/LegacyUI/Toolbar.cpp
@@ -18,8 +18,8 @@
 #include "Spire/Charting/ChartWindow.hpp"
 #include "Spire/Dashboard/DashboardWindow.hpp"
 #include "Spire/Dashboard/DashboardModelSchema.hpp"
-#include "Spire/KeyBindings/SimplifiedKeyBindingsDialog.hpp"
-#include "Spire/KeyBindings/TaskKeyBindingsDialog.hpp"
+#include "Spire/LegacyKeyBindings/SimplifiedKeyBindingsDialog.hpp"
+#include "Spire/LegacyKeyBindings/TaskKeyBindingsDialog.hpp"
 #include "Spire/LegacyUI/CanvasWindow.hpp"
 #include "Spire/LegacyUI/FunctionalAction.hpp"
 #include "Spire/LegacyUI/ImportExportDialog.hpp"
