+++ Applications/Spire/Source/Spire/main.cpp
@@ -18,14 +18,14 @@
 #include "Spire/Blotter/BlotterWindow.hpp"
 #include "Spire/BookView/BookViewWindow.hpp"
 #include "Spire/Dashboard/SavedDashboards.hpp"
-#include "Spire/OrderImbalanceIndicator/OrderImbalanceIndicatorProperties.hpp"
-#include "Spire/KeyBindings/HotkeyOverride.hpp"
+#include "Spire/LegacyKeyBindings/HotkeyOverride.hpp"
 #include "Spire/LegacyUI/CustomQtVariants.hpp"
 #include "Spire/LegacyUI/Toolbar.hpp"
 #include "Spire/LegacyUI/UserProfile.hpp"
 #include "Spire/LegacyUI/WindowSettings.hpp"
 #include "Spire/Login/LoginController.hpp"
 #include "Spire/Login/LoginException.hpp"
+#include "Spire/OrderImbalanceIndicator/OrderImbalanceIndicatorProperties.hpp"
 #include "Spire/PortfolioViewer/PortfolioViewerProperties.hpp"
 #include "Spire/RiskTimer/RiskTimerMonitor.hpp"
 #include "Spire/Spire/Resources.hpp"
